Attendance cutover to att_v2
- Home / now renders the v2 check-in (shared CheckinV2 component, deep-blue);
  staff write directly to att_v2 via checkin_v2/checkout_v2 RPCs.
- /attendance/checkin reuses the same component.
- Nav: staff see Check-in only; admin keeps Attendance tab + Niagawan +
  Employees + Payroll (v3) + Records. Removed stale old /today /report
  /offday top-level links and the Payroll v2 link.
- (DB done separately: geofence reverted to real radius_m; mirror kept as a
  safety no-op for now.)
